requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.6
Pillow==9.5.0
poe-api==0.3.1
PySocks==1.7.1
python-pptx==0.6.21
python-socks==2.3.0
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
websocket-client==1.5.1
XlsxWriter==3.1.0
yarl==1.9.2
